refactor: Console -> ConsoleInput, ConsoleOutput 분리
