+++ src/main/java/Application.java
@@ -1,6 +1,10 @@
+import io.Input;
+import io.Output;
+
 public class Application {
     public static void main(String[] args) {
-        Console console = new Console();
-        new Calculator(console, console).run();
+        Output output = new ConsoleOutput();
+        Input input = new ConsoleInput();
+        new Calculator(input, output).run();
     }
 }
